Remove data_array_view and simplify conversions to numpy and cupy
