+++ python/cudf/cudf/core/groupby/groupby.py
@@ -1593,7 +1593,7 @@ class GroupBy(Serializable, Reducible, Scannable):
                         [plc.types.NullOrder.AFTER],
                     )
                     indices = ColumnBase.from_pylibcudf(plc_table.columns()[0])
-                indices = indices.data_array_view(mode="read")
+                indices = indices.values
             # Which indices are we going to want?
             want = np.arange(samples_per_group.sum(), dtype=SIZE_TYPE_DTYPE)
             scan = np.empty_like(samples_per_group)
