+++ python/cudf/cudf/utils/queryutils.py
@@ -224,7 +224,7 @@ def query_execute(df, expr, callenv):
             "query only supports numeric, datetime, timedelta, or bool dtypes."
         )
 
-    colarrays = [col.data_array_view(mode="read") for col in colarrays]
+    colarrays = [col.values for col in colarrays]
 
     kernel = compiled["kernel"]
     # process env args
